refactor: user 구조 변경 #5
