+++ mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/account/Account.java
@@ -1,6 +1,6 @@
 package com.ssafy.mockstockinvestment.account;
 
-import com.ssafy.mockstockinvestment.user.Student;
+import com.ssafy.mockstockinvestment.user.domain.Student;
 import com.sun.istack.NotNull;
 import lombok.*;
