+++ mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/account/AccountRepository.java
@@ -1,10 +1,8 @@
 package com.ssafy.mockstockinvestment.account;
 
-import com.ssafy.mockstockinvestment.user.Student;
+import com.ssafy.mockstockinvestment.user.domain.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-
-import java.sql.Timestamp;
 
 public interface AccountRepository extends JpaRepository<Account, Long> {
     @Query(value = "select a from accounts a where a.studentIdFK=?1 and a.createdAt=?2 and a.assetType=?3")
