--- mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/StudentRepository.java
@@ -1,6 +1,0 @@
-package com.ssafy.mockstockinvestment.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudentRepository extends JpaRepository<Student, String> {
-}
