--- mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/UserEnum.java
@@ -1,5 +1,0 @@
-package com.ssafy.mockstockinvestment.user;
-
-public enum UserEnum {
-    관리자, 학생
-}
