--- mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/UserRepository.java
@@ -1,8 +1,0 @@
-package com.ssafy.mockstockinvestment.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User,Long>{
-}
