--- mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/UserService.java
@@ -1,6 +1,0 @@
-package com.ssafy.mockstockinvestment.user;
-
-
-public interface UserService {
-
-}
