--- mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/UserServiceImpl.java
@@ -1,15 +1,0 @@
-package com.ssafy.mockstockinvestment.user;
-
-import org.springframework.stereotype.Service;
-
-import javax.transaction.Transactional;
-
-@Service
-@Transactional
-public class UserServiceImpl implements UserService{
-    private final UserRepository userRepository;
-
-    public UserServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-}
