+++ mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/controller/UserController.java
@@ -1,7 +1,7 @@
-package com.ssafy.mockstockinvestment.user;
+package com.ssafy.mockstockinvestment.user.controller;
 
+import com.ssafy.mockstockinvestment.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
