+++ mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/domain/JwtTokenProvider.java
@@ -1,0 +1,4 @@
+package com.ssafy.mockstockinvestment.user.domain;
+
+public class JwtTokenProvider {
+}
