+++ mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/domain/Student.java
@@ -1,4 +1,4 @@
-package com.ssafy.mockstockinvestment.user;
+package com.ssafy.mockstockinvestment.user.domain;
 
 import com.ssafy.mockstockinvestment.project.Project;
 import com.ssafy.mockstockinvestment.stock.Stock;
