+++ mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/domain/UserEnum.java
@@ -1,0 +1,5 @@
+package com.ssafy.mockstockinvestment.user.domain;
+
+public enum UserEnum {
+    관리자, 학생
+}
