+++ mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/domain/repository/StudentRepository.java
@@ -1,0 +1,4 @@
+package com.ssafy.mockstockinvestment.user.domain.repository;
+
+public interface StudentRepository {
+}
