+++ mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/dto/request/LoginRequest.java
@@ -1,0 +1,4 @@
+package com.ssafy.mockstockinvestment.user.dto.request;
+
+public class LoginRequest {
+}
