+++ mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/dto/response/TokenRequest.java
@@ -1,0 +1,4 @@
+package com.ssafy.mockstockinvestment.user.dto.response;
+
+public class TokenRequest {
+}
