+++ mock-stock-investment/src/main/java/com/ssafy/mockstockinvestment/user/service/UserService.java
@@ -1,0 +1,4 @@
+package com.ssafy.mockstockinvestment.user.service;
+
+public class UserService {
+}
